Add JaCoCo to Java Coverage Example (#14)
* Testing new Instrumentation

* Trying to get Coverage Reports working

* Trying a new approach

* Trying Again

* Re-adding Emma for Comparison

* Trying Again...

* Try Try Try Again

* Finally, It (Kinda) Works

Instead of doing Instrumentation the way Emma does, I am using on-demand instrumentation. It's better documented and (kinda) works. The corresponding changes involving libraries and filenames have not been added to Submitty yet, but I have tested this on my computer and it seems to work. I have not been able to test TestRunner specifically, but I did test running student tests through `org.junit.runner.JUnitCore` and it seemed to work fine.

* Other Fixes

This is so the new descriptions are accurate

* Trying to View Results

* modified the java coverage test to use show use of jacoco and emma

* replace tabs with spaces