[Testing:Autograding] move course.yml from Submitty.Submitty (#45)
This PR adds a `course_config.yml` file to the root of the repo. It is
copied directly from `.setup/data/courses/tutorial.yml` in
Submitty/Submitty. This will be part of a refactor to move
repo-dependent course yml files into their respective repos. For now,
this PR does not do anything.